+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - Add `GlDisplay::version_string` to help with logging the display information.
 - Rename `NotCurrentGlContext::treat_as_current` to `NotCurrentGlContext::treat_as_possibly_current`.
 - Rename `Display::from_raw` to `Display::new`.
+- Added `GlDisplay::supported_features` to allow checking for extensions support beforehand.
+- **Breaking:** renamed `ReleaseBehaviour` to `ReleaseBehavior`.
 
 # Version 0.30.0-beta.2 (2022-09-03)